+++ lib/src/cupertino/containers/colored_box.dart
@@ -3,41 +3,41 @@ import 'package:flutter/cupertino.dart';
 import 'package:exui/src/widget/containers/colored_box.dart';
 
 extension CupertinoBoxColorsExtension on Widget {
-  Widget redBox() => coloredBox(CupertinoColors.systemRed);
+  ColoredBox redBox() => coloredBox(CupertinoColors.systemRed);
 
-  Widget destructiveRedBox() => coloredBox(CupertinoColors.destructiveRed);
+  ColoredBox destructiveRedBox() => coloredBox(CupertinoColors.destructiveRed);
 
-  Widget greenBox() => coloredBox(CupertinoColors.systemGreen);
+  ColoredBox greenBox() => coloredBox(CupertinoColors.systemGreen);
 
-  Widget activeGreenBox() => coloredBox(CupertinoColors.activeGreen);
+  ColoredBox activeGreenBox() => coloredBox(CupertinoColors.activeGreen);
 
-  Widget blueBox() => coloredBox(CupertinoColors.systemBlue);
+  ColoredBox blueBox() => coloredBox(CupertinoColors.systemBlue);
 
-  Widget activeBlueBox() => coloredBox(CupertinoColors.activeBlue);
+  ColoredBox activeBlueBox() => coloredBox(CupertinoColors.activeBlue);
 
-  Widget yellowBox() => coloredBox(CupertinoColors.systemYellow);
+  ColoredBox yellowBox() => coloredBox(CupertinoColors.systemYellow);
 
-  Widget orangeBox() => coloredBox(CupertinoColors.systemOrange);
+  ColoredBox orangeBox() => coloredBox(CupertinoColors.systemOrange);
 
-  Widget orangeAccentBox() => coloredBox(CupertinoColors.activeOrange);
+  ColoredBox orangeAccentBox() => coloredBox(CupertinoColors.activeOrange);
 
-  Widget purpleBox() => coloredBox(CupertinoColors.systemPurple);
+  ColoredBox purpleBox() => coloredBox(CupertinoColors.systemPurple);
 
-  Widget pinkBox() => coloredBox(CupertinoColors.systemPink);
+  ColoredBox pinkBox() => coloredBox(CupertinoColors.systemPink);
 
-  Widget brownBox() => coloredBox(CupertinoColors.systemBrown);
+  ColoredBox brownBox() => coloredBox(CupertinoColors.systemBrown);
 
-  Widget tealBox() => coloredBox(CupertinoColors.systemTeal);
+  ColoredBox tealBox() => coloredBox(CupertinoColors.systemTeal);
 
-  Widget cyanBox() => coloredBox(CupertinoColors.systemCyan);
+  ColoredBox cyanBox() => coloredBox(CupertinoColors.systemCyan);
 
-  Widget greyBox() => coloredBox(CupertinoColors.systemGrey);
+  ColoredBox greyBox() => coloredBox(CupertinoColors.systemGrey);
 
-  Widget blackBox() => coloredBox(CupertinoColors.black);
+  ColoredBox blackBox() => coloredBox(CupertinoColors.black);
 
-  Widget whiteBox() => coloredBox(CupertinoColors.white);
+  ColoredBox whiteBox() => coloredBox(CupertinoColors.white);
 
-  Widget darkGrayBox() => coloredBox(CupertinoColors.darkBackgroundGray);
+  ColoredBox darkGrayBox() => coloredBox(CupertinoColors.darkBackgroundGray);
 
-  Widget lightGrayBox() => coloredBox(CupertinoColors.lightBackgroundGray);
+  ColoredBox lightGrayBox() => coloredBox(CupertinoColors.lightBackgroundGray);
 }