+++ lib/src/widget/clip/clip.dart
@@ -1,4 +1,3 @@
 export 'clip_path.dart';
 export 'clip_rect.dart';
 export 'squircle.dart';
-export 'superellipse.dart';