+++ lib/src/widget/clip/clip_rect.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/widgets.dart';
 
 extension ClipRectWidgetExtensions on Widget {
-  Widget clipRRect({
+  ClipRRect clipRRect({
     BorderRadiusGeometry borderRadius = BorderRadius.zero,
     CustomClipper<RRect>? clipper,
     Clip clipBehavior = Clip.antiAlias,
@@ -14,7 +14,18 @@ extension ClipRectWidgetExtensions on Widget {
     );
   }
 
-  Widget clipOval({
+  ClipRRect clipCircular([
+    double radius = 12,
+    Clip clipBehavior = Clip.antiAlias,
+  ]) {
+    return ClipRRect(
+      borderRadius: BorderRadius.circular(radius),
+      clipBehavior: clipBehavior,
+      child: this,
+    );
+  }
+
+  ClipOval clipOval({
     CustomClipper<Rect>? clipper,
     Clip clipBehavior = Clip.antiAlias,
   }) {
@@ -25,7 +36,7 @@ extension ClipRectWidgetExtensions on Widget {
     );
   }
 
-  Widget clipCircle({
+  ClipOval clipCircle({
     CustomClipper<Rect>? clipper,
     Clip clipBehavior = Clip.antiAlias,
   }) =>