+++ lib/src/widget/style/opacity.dart
@@ -1,16 +1,22 @@
 import 'package:flutter/widgets.dart';
 
 extension OpacityWidgetExtension on Widget {
-  Widget opacity(double opacity, {bool alwaysIncludeSemantics = false}) =>
+  Opacity opacity(double opacity, {bool alwaysIncludeSemantics = false}) =>
       Opacity(
         opacity: opacity,
         alwaysIncludeSemantics: alwaysIncludeSemantics,
         child: this,
       );
 
-  Widget opacityPercent(double percent) => opacity(percent / 100);
+  Opacity opacityPercent(double percent) => opacity(percent / 100);
 
-  Widget opacityHalf() => opacity(0.5);
+  Opacity opacityHalf() => opacity(0.5);
 
-  Widget opacityQuarter() => opacity(0.25);
+  Opacity opacityQuarter() => opacity(0.25);
+
+  Opacity opacityZero() => opacity(0);
+
+  Opacity opacityTransparent() => opacityZero();
+
+  Opacity opacityInvisible() => opacityZero();
 }